Extract things that have nothing to do with graph building
