+++ src/main/java/minijava/ir/IREmitter.java
@@ -31,7 +31,7 @@ public class IREmitter
         BlockStatement.Visitor<Void> {
 
   private static final Type INT_TYPE;
-  private static final Type BOOLEAN_TYPE;
+  static final Type BOOLEAN_TYPE;
   private static final Type PTR_TYPE;
   private static final MethodType CALLOC_TYPE;
   private static final Entity CALLOC;
@@ -96,39 +96,11 @@ public class IREmitter
     classTypes.clear();
     methods.clear();
     fields.clear();
-    for (Class decl : that.declarations) {
-      ClassType classType = new ClassType(decl.name());
-      classTypes.put(decl, classType);
-      for (Field f : decl.fields) {
-        fields.put(f, createEntity(f));
-      }
-      if (decl.fields.size() == 0) {
-        // We have to prevent class types of length 0, so we insert an unreachable field.
-        String unusableFieldName = "0padding"; // unusable because of digit prefix
-        Entity fieldEnt = new Entity(classType, unusableFieldName, BOOLEAN_TYPE);
-        fieldEnt.setLdIdent(NameMangler.mangleInstanceFieldName(decl.name(), unusableFieldName));
-      }
-      for (Method m : decl.methods) {
-        methods.put(m, createEntity(m));
-      }
-      //System.out.println("# " + classType);
-      //System.out.println(classType.getSize());
-      classType.layoutFields();
-      classType.finishLayout();
-      //System.out.println(classType.getSize());
-    }
+    that.acceptVisitor(new Collector(classTypes, fields, methods));
     for (Class klass : that.declarations) {
       klass.methods.forEach(this::emitBody);
     }
     return null;
-  }
-
-  private Entity createEntity(Field f) {
-    Type type = getStorageType(f.type);
-    ClassType definingClass = classTypes.get(f.definingClass.def);
-    Entity fieldEnt = new Entity(definingClass, f.name(), type);
-    fieldEnt.setLdIdent(NameMangler.mangleInstanceFieldName(definingClass.getName(), f.name()));
-    return fieldEnt;
   }
 
   /**
@@ -139,49 +111,9 @@ public class IREmitter
    * <p>So, for value types just return their value, for reference types return a pointer.
    */
   @NotNull
-  private PrimitiveType getStorageType(minijava.ast.Type type) {
+  static PrimitiveType getStorageType(minijava.ast.Type type) {
     // TODO shouldn't we reuse these types?
     return new PrimitiveType(storageModeForType(type));
-  }
-
-  /**
-   * This will *not* go through the body of the method, just analyze stuff that is needed for
-   * constructing an entity.
-   */
-  private Entity createEntity(Method m) {
-    if (m.isStatic) {
-      return addMainMethodDecl(m);
-    }
-    ClassType definingClass = classTypes.get(m.definingClass.def);
-    ArrayList<Type> parameterTypes = new ArrayList<>();
-
-    // Add the this pointer. It's always parameter 0, so access will be trivial.
-    parameterTypes.add(ptrTo(definingClass));
-    for (LocalVariable p : m.parameters) {
-      // In the body, we need to refer to local variables by index, so we save that mapping.
-      parameterTypes.add(getStorageType(p.type));
-    }
-
-    // The visitor returns null if that.returnType was void.
-    Type[] returnTypes = {};
-    if (!m.returnType.equals(minijava.ast.Type.VOID)) {
-      returnTypes = new Type[] {getStorageType(m.returnType)};
-    }
-
-    Type methodType = new MethodType(parameterTypes.toArray(new Type[0]), returnTypes);
-
-    // Set the mangled name
-    Entity methodEnt = new Entity(definingClass, m.name(), methodType);
-    methodEnt.setLdIdent(NameMangler.mangleMethodName(definingClass.getName(), m.name()));
-    return methodEnt;
-  }
-
-  private Entity addMainMethodDecl(Method m) {
-    MethodType type = new MethodType(0, 0);
-    SegmentType global = Program.getGlobalType();
-    Entity mainEnt = new Entity(global, "main", type);
-    mainEnt.setLdIdent(NameMangler.mangledMainMethodName());
-    return mainEnt;
   }
 
   private void emitBody(Method m) {
@@ -312,7 +244,7 @@ public class IREmitter
     return classTypes.get(that);
   }
 
-  private Mode storageModeForType(minijava.ast.Type type) {
+  static Mode storageModeForType(minijava.ast.Type type) {
     if (type.dimension > 0) {
       return Mode.getP();
     }
@@ -325,10 +257,6 @@ public class IREmitter
       default:
         return Mode.getP();
     }
-  }
-
-  private PointerType ptrTo(Type type) {
-    return new PointerType(type);
   }
 
   @Override
